% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{annx}
\alias{annx}
\title{EPV of Annuity}
\usage{
annx(x = gl.g(x), s = 0, i = gl.g(i), m = 1, n = gl.g(w) - x, c = 0,
  e = 1, mt = 1)
}
\arguments{
\item{x}{the current age}

\item{s}{the select used so far}

\item{i}{the interest rate}

\item{m}{the compounding frequency}

\item{n}{the length of the term}

\item{c}{indicator of continuous (1 if continuous)}

\item{e}{indicator of endowment (NOTE of an annuity should always be 1)}

\item{mt}{the moment of the insurance}
}
\description{
Calculates the Expected Presented Value of various annuities
}
\details{
By default calculates the first moment of discrete, whole life annuity due
}

