% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{Ax}
\alias{Ax}
\title{EPV of Insurance}
\usage{
Ax(x = global$x, s = 0, i = global$i, m = 1, n = global$w - x,
  c = 0, e = 0, mt = 1)
}
\arguments{
\item{x}{the current age}

\item{s}{the select used so far}

\item{i}{the interest rate}

\item{m}{the compounding frequency}

\item{n}{the length of the term}

\item{c}{indicator of continuous (1 if continuous)}

\item{e}{indicator of endowment (1 if endowment)}

\item{mt}{the moment of the insurance}
}
\description{
Calculates the Expected Presented Value of various insurances
}
\details{
Default: first moment of discrete, whole life insurance
}

