% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{cfm}
\alias{cfm}
\title{Change Survival Model to CFM}
\usage{
cfm(mu = 0.04, delta = log(1 + gl.g(i)), w = 1000)
}
\arguments{
\item{mu}{the force of mortality}

\item{delta}{the force of interest}

\item{w}{the arbitrarily large limiting age}
}
\description{
Changes parameters and force of mortality function
             to use constant force of mortality
}
\details{
To revert to makehams use makehams()
}

