% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{createLifeTable}
\alias{createLifeTable}
\title{Create Ultimate Select Life Table}
\usage{
createLifeTable(x = gl.g(x), w = gl.g(w), radix = gl.g(radix),
  d = gl.g(d))
}
\arguments{
\item{x}{the starting age for the life table}

\item{w}{the limiting age}

\item{radix}{the number of individuals aged x}

\item{d}{the select period}
}
\description{
Creates a life table based on the select period, radix and Makeham model parameters
}
\details{
See Appendix Tables of DHW 2nd edition
}

