% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{d}
\alias{d}
\title{Nominal rate of discount}
\usage{
d(i = gl.g(i), m = 1)
}
\arguments{
\item{i}{the effective annual interest rate}

\item{m}{the number of compounding periods}
}
\description{
Calculates the interest rate used for annuity due's
}
\details{
Uses the relation (1-d) = (1-d(m)/m)^(-m)
}
\keyword{internal}

