% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{dannx}
\alias{dannx}
\title{EPV of a Deferred Annuity}
\usage{
dannx(u = 0, x = gl.g(x), s = 0, i = gl.g(i), m = 1, n = gl.g(w) -
  x, c = 0, e = 1, mt = 1)
}
\arguments{
\item{u}{the length of the deferral period}

\item{x}{the current age}

\item{s}{the select used so far}

\item{i}{the interest rate}

\item{m}{the compounding frequency}

\item{n}{the length of the term}

\item{c}{indicator of continuous (1 if continuous)}

\item{e}{indicator of endowment (NOTE of an annuity should always be 1)}

\item{mt}{the moment of the insurance}
}
\description{
Calculates the expected present value of a deferred annuity
}
\details{
By default calculates the first moment of discrete, whole life annuity due.
         Also, this function is not reliable when n < m.
}

