% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{dAx}
\alias{dAx}
\title{EPV of a Deferred Insurance}
\usage{
dAx(u = 0, x = gl.g(x), s = 0, i = gl.g(i), m = 1, n = gl.g(w) - x,
  c = 0, e = 0, mt = 1)
}
\arguments{
\item{u}{the length of the deferral period}

\item{x}{the current age}

\item{s}{the select used so far}

\item{i}{the interest rate}

\item{m}{the compounding frequency}

\item{n}{the length of the term}

\item{c}{indicator of continuous (1 if continuous)}

\item{e}{indicator of endowment (1 if endowment)}

\item{mt}{the moment of the insurance}
}
\description{
Calculates the expected present value of a deferred insurance
}
\details{
By default calculates first moment of discrete, whole life insurance.
         Also, this function is not reliable when n < m.
         u can be a vector with length > 1
         x can be a vector with length > 1
         s can be a vector with length > 1
         n can be a vector with length > 1
}

