% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{demoivres}
\alias{demoivres}
\title{Change Survival Model to DeMoivre's}
\usage{
demoivres(w = 100, delta = log(1 + gl.g(i)))
}
\arguments{
\item{w}{the limiting age}

\item{delta}{the force of interest}
}
\description{
Change Survival Model to DeMoivre's
}
\details{
Changes parameters and force of interest function to Uniform model
}

