% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{gl.a}
\alias{gl.a}
\title{Assign global varaiable value}
\usage{
gl.a(var, val = NA)
}
\arguments{
\item{var}{the variable name to assign (object name, not a string)}

\item{val}{the value to assign to the variable}
}
\description{
Assigns or overrides an existing variable in the global variables environment
}
\details{
by default, the value NA is assigned when unspecified
}
\keyword{internal}

