% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{gl.g}
\alias{gl.g}
\title{Retrives global variable value}
\usage{
gl.g(var)
}
\arguments{
\item{var}{the variable name to get the value of}
}
\description{
Retrives the value of a variable defined
             in the global variables environment
}
\details{
the var argument should be an object name, not a string
}
\keyword{internal}

