% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\docType{data}
\name{gl}
\alias{gl}
\title{Global variables Environment}
\format{\preformatted{<environment: 0x00000000147711b0> 
}}
\usage{
gl
}
\description{
Model variables are stored are retrieved from this environment,
             such as the interest rate and limiting age
}
\details{
To see variables assigned by default use ls(envir=gl)
}
\keyword{internal}

