% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\docType{data}
\name{global}
\alias{global}
\title{globalParameters}
\format{\preformatted{List of 8
 $ A    : num 0.00022
 $ B    : num 2.7e-06
 $ c    : num 1.12
 $ d    : num 2
 $ x    : num 20
 $ w    : num 131
 $ radix: num 1e+05
 $ i    : num 0.05
}}
\usage{
global
}
\description{
Contains all of the global parameters used by the Makeham Model.
             "A","B","c" are makeham model params, "d" is select period, "x" is default age,
             "w" is limiting age, "radix" is used as initial survivors for life tables,
             "i" is the effective annual interest rate.
}
\keyword{datasets}

