% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{makehams}
\alias{makehams}
\title{Change Survival to Makeham's}
\usage{
makehams(A = 0.00022, B = 2.7e-06, c = 1.124, d = 2, x = 20,
  w = 131, radix = 1e+05, i = 0.05)
}
\arguments{
\item{A}{model parameter}

\item{B}{model parameter}

\item{c}{model parameter}

\item{d}{select period}

\item{x}{the default age}

\item{w}{the limiting age}

\item{radix}{the number of starting individuals in life table}

\item{i}{the effective annual interest rate}
}
\description{
Change Survival to Makeham's
}
\details{
Reverts the survival model back to Makeham's law with
         default parameters
}

