% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{tEx}
\alias{tEx}
\title{Actuarial Present Value Factor}
\usage{
tEx(t, x = global$x, s = 0, i = global$i, mt = 1)
}
\arguments{
\item{t}{the years from x}

\item{x}{the current age}

\item{s}{the select used so far}

\item{i}{the interest rate}

\item{mt}{the moment of the insurance}
}
\description{
Calculates the Expected Present value of a pure endowment insurance
}
\details{
Alternative actuarial "A" notation is also used for tEx
}

