% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{thV}
\alias{thV}
\title{Benefit Reserve}
\usage{
thV(t = 0, h = 1, x = gl.g(x), tV = 0, Pt = function(t) {     t^0 *
  gl.g(pi) }, deltat = function(t) {     t^0 * log(1 + gl.g(i)) },
  bt = function(t) {     t^0 }, ut = function(t) {     uxt(t, x) },
  s = 0.01)
}
\arguments{
\item{t}{the time for which the reserve is known}

\item{h}{the the time from t for which the reserve should be calculated}

\item{x}{the age of the person for which the reserve is being calculated}

\item{tV}{the value of the reserve at time t}

\item{Pt}{the premium as a function of t}

\item{deltat}{the force of interest as a function of t}

\item{bt}{the death benefit payable immediately at the time of
death as a function of t}

\item{ut}{the force of mortality as a function of t}

\item{s}{the step to use in Euler's method}
}
\description{
Uses Euler's method to solve Thiele's differential equation to
             approximate the value of t+hV
}
\details{
This function does not take into account expenses
}

