% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{tpx}
\alias{tpx}
\title{Survival Function}
\usage{
tpx(t, x = gl.g(x), s = 0, uxt = gl.g(uxt))
}
\arguments{
\item{t}{the number of years to survive}

\item{x}{the current age}

\item{s}{select already used}

\item{uxt}{the force of mortality (can be used to override the default force of mortality)}
}
\description{
Probability that x survives t years given survival to age x
}
\details{
Uses a default select period of 2 (for makeham's law)
}

