% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{tqx}
\alias{tqx}
\title{CDF of Future Lifetime}
\usage{
tqx(t, x = global$x, s = 0)
}
\arguments{
\item{t}{the number of years before death}

\item{x}{the current age}
}
\description{
Probability that x dies in the next t years, given survival to age x
}
\details{
Calcualted as 1 - tpx(t,x)
}

