% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{udeferredtqx}
\alias{udeferredtqx}
\title{Deferred CDF of Future Lifetime}
\usage{
udeferredtqx(u, t = 1, x = global$x, s = 0)
}
\arguments{
\item{u}{the number of years to survive}

\item{t}{the number of years to death within}

\item{x}{the current age}

\item{s}{the select used}
}
\description{
Probability of surviving u years and dying in the next t years
}
\details{
Can be calculated by splitting the CDF. Use tpx(u,x) - tpx(u+t,x)
}

