% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{updateParams}
\alias{updateParams}
\title{updateParams}
\usage{
updateParams(paramNames, paramVals)
}
\arguments{
\item{paramNames}{a string vector containing the names of variables to be updated}

\item{paramVals}{a numeric vector containing the values of variables to be updated}
}
\description{
Update the Makeham Model Parameters
}
\details{
"A", "B", "c" are Makeham Model Parameters. "radix" can be updated for life tables.
         "d" is the select period. "x" is the default age when not specified. "w" is the limiting age.
}
\examples{
updateParams(c("A","B"), c(0.00022,2.7e-06))
}

