% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{uxt}
\alias{uxt}
\title{Force of Mortality}
\usage{
uxt(t, x = gl.g(x), s = 0, d = gl.g(d), A = gl.g(A), B = gl.g(B),
  c = gl.g(c))
}
\arguments{
\item{t}{the years after age x}

\item{x}{the current age}

\item{s}{select already used}

\item{d}{the select period}

\item{A}{Makeham model parameter}

\item{B}{Makeham model parameter}

\item{c}{Makeham model parameter}
}
\description{
The select force of mortality, u[x]+s = 0.9^(2-s) ux+s
             where the force of mortality is ux+s = A + Bc^(x+t)
}

