% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{v}
\alias{v}
\title{Present Value Factor}
\usage{
v(i = global$i, n = 1, delta = log(1 + i))
}
\arguments{
\item{i}{the effective annual interest rate}

\item{n}{the number of years to apply discounting}

\item{delta}{the force of interest}

\item{The}{force of interest is internally derived from the effective annual interest rate}
}
\description{
Calculates the present value of a cash flow
}

