% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{Ax}
\alias{Ax}
\title{EPV of Insurance}
\usage{
Ax(x = gl.g(x), s = 0, i = gl.g(i), m = 1, n = gl.g(w) - x, c = 0,
  e = 0, mt = 1, h = function(t) t^0)
}
\arguments{
\item{x}{the current age}

\item{s}{the select used so far}

\item{i}{the interest rate}

\item{m}{the compounding frequency}

\item{n}{the length of the term}

\item{c}{indicator of continuous (1 if continuous)}

\item{e}{indicator of endowment (1 if endowment)}

\item{mt}{the moment of the insurance}
}
\description{
Calculates the Expected Presented Value of various insurances
}
\details{
By default calculates first moment of discrete, whole life insurance.
         Also, this function is not reliable when n < m.
         x can be a vector with length > 1
         s can be a vector with length > 1
         n can be a vector with length > 1
}

