% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{createInsuranceTable}
\alias{createInsuranceTable}
\title{Create Insurance Table}
\usage{
createInsuranceTable(x = gl.g(x), w = gl.g(w), d = gl.g(d), n = 5,
  i = gl.g(i), mt = 1)
}
\arguments{
\item{x}{the starting age}

\item{w}{the limiting age}

\item{d}{the select period}

\item{n}{pure endowment period}

\item{i}{the interest rate}

\item{mt}{the moment t calculate}
}
\description{
Creates a table containing EPV's of whole life insurances (discrete)
}
\details{
Computes life table using recursion
}

