% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makehams.R
\name{tpx}
\alias{tpx}
\title{Survival Function}
\usage{
tpx(t, x = global$x, s = 0)
}
\arguments{
\item{t}{the number of years to survive}

\item{x}{the current age}

\item{s}{select already used}
}
\description{
Probability that x survives t years given survival to age x
}
\details{
Uses a default select period of 2, but this can be changed in global$d
}

