% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{tpx}
\alias{tpx}
\title{Survival Function}
\usage{
tpx(t, x = gl.g(x), s = 0, uxt = gl.g(uxt), addtox = FALSE)
}
\arguments{
\item{t}{the number of years to survive}

\item{x}{the current age}

\item{s}{select already used}

\item{uxt}{the force of mortality (can be used to override the
default force of mortality)}

\item{addtox}{indicate whether s implies that age is x+s}
}
\description{
Probability that x survives t years given survival to age x
}
\details{
Uses a default select period of 2 (for makeham's law).
         t can be a vector with length > 1
         x can be a vector with length > 1
         s can be a vector with length > 1
}

