% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{tqx}
\alias{tqx}
\title{CDF of Future Lifetime}
\usage{
tqx(t, x = gl.g(x), s = 0, uxt = gl.g(uxt), addtox = FALSE)
}
\arguments{
\item{t}{the number of years before death}

\item{x}{the current age}

\item{s}{select already used}

\item{uxt}{the force of mortality (can be used to override the
default force of mortality)}

\item{addtox}{indicate whether s implies age is x+s}
}
\description{
Probability that x dies in the next t years, given survival to age x
}
\details{
Calcualted as 1 - tpx(t,x)
}

