% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makehams.R
\name{udeferredtqx}
\alias{udeferredtqx}
\title{Deferred CDF of Future Lifetime}
\usage{
udeferredtqx(u, t = 1, x = gl.g(x), s = 0, addtox = FALSE)
}
\arguments{
\item{u}{the number of years to survive}

\item{t}{the number of years to death within}

\item{x}{the current age}

\item{s}{the select used}

\item{addtox}{indicate whether s implies that age is x+s}
}
\description{
Probability of surviving u years and dying in the next t years
}
\details{
Can be calculated by splitting the CDF. Use tpx(u,x) - tpx(u+t,x) since it
         tends to be more accurate than tpx * tqx method.
         u can be a vector with length > 1
         t can be a vector with length > 1
         x can be a vector with length > 1
         s can be a vector with length > 1
}

